Ui screen logic separated
